Refactor time_base and time_scale
Instead of associating them with each chunk, make them attributes of the
class.

This moves away a little from having each xr.DataArray be
self-contained. The advantage is that now these values are known to the
constructor (in future this could be used in e.g. VDIFEncoder to ensure
that there are an integer number of frames per second etc), and there is
no possibility of them being set inconsistently between chunks.