Fix VDIF encoding when sync time is not on a word boundary
VDIFEncoder2Bit requires time_bias to be a multiple of SAMPLES_PER_WORD,
but this was not guaranteed. Since chunks are also required to be
aligned to frame boundaries, this can only be guaranteed by adjusting
the time_base to fall on a word boundary. In practice, it's simplest to
align it to a second boundary.

This makes --time-correction usable in mk_vlbi_resample.