Do ConvertPolarisation before Resample
I'd moved it afterwards thinking it would be beneficial to do it at the
lower sampling rate after resampling, but the resampling also splits out
the complex signal to real sidebands, and that doesn't commute with a
(complex) Jones matrix.

Just do ConvertPolarisation before Resample. It might be fractionally
more performant to do it in the middle (between upfirdn and
split_sidebands) but it doesn't seem to be a performance bottleneck.