Use asyncio.Event instead of cuda events
cupy's launch_host_func makes it possible to get asynchronous callback
when queued work is complete. That's a nicer model for integration with
asyncio than CUDA events, which don't have a properly non-blocking
mechanism for notifications. One can use a worker thread to wait, but
cupy doesn't seem to support cudaSetDeviceFlags, so it is likely to end
up spinning rather than yielding.